refactor: move main components to main folder
