+++ src/component/main/SplitPaneWrapper.tsx
@@ -1,6 +1,6 @@
 import { SplitPane, SplitPaneSize } from 'react-science/ui';
 
-import { usePreferences } from './context/PreferencesContext';
+import { usePreferences } from '../context/PreferencesContext';
 
 export function SplitPaneWrapper({ children }) {
   const { current, dispatch } = usePreferences();
