+++ src/component/main/index.ts
@@ -2,4 +2,4 @@ export * from './types';
 
 export { serializeNmriumState as serializeNMRiumState } from 'nmr-load-save';
 
-export * from '../NMRium';
+export * from './NMRium';
